+++ README.md
@@ -2,5 +2,4 @@
 
 ## Démarrer le projet
 
-Rien à installer ici, il suffit d'ouvrir le fichier `index.html`.
-
+Dépendance SASS & WebPack & Babel & PostCSS
