+++ index.html
@@ -1,7 +1,6 @@
 <!DOCTYPE html>
 <html>
     <head>
-        <link href="css/style.css" rel="stylesheet" type="text/css">
         <meta charset="utf-8">
         <link rel="icon" type="image/png" href="assets/favicon.png">
         <link href="https://fonts.googleapis.com/css?family=DM+Sans&display=swap" rel="stylesheet">
@@ -9,7 +8,7 @@
     </head>
     <body>
         <header>
-            <img src="assets/images/logo.png" class="logo">
+            <img src="./assets/images/logo.png" class="logo">
             <h1>Nos photographes</h1>
         </header>
         <main id="main">
@@ -17,9 +16,10 @@
         </main>
         <footer></footer>
     </body>
-    <script src="scripts/utils/fetch.js"></script>
+    <script src="./dist/index.js"></script>
+    <!-- <script src="scripts/utils/fetch.js"></script>
     <script src="scripts/utils/dom.js"></script>
     <script src="scripts/factories/photographerFactory.js"></script>
     <script src="scripts/pages/sharedFunction.js"></script>
-    <script src="scripts/pages/index.js"></script>
+    <script src="scripts/pages/index.js"></script> -->
 </html>
