+++ photographer.html
@@ -1,7 +1,6 @@
 <!DOCTYPE html>
 <html>
     <head>
-        <link href="css/style.css" rel="stylesheet" type="text/css">
         <meta charset="utf-8">
         <link rel="icon" type="image/png" href="assets/favicon.png">
         <link href="https://fonts.googleapis.com/css?family=DM+Sans&display=swap" rel="stylesheet">
@@ -9,7 +8,7 @@
     </head>
     <body>
         <header>
-            <img src="assets/images/logo.png" class="logo_photographer">
+            <img src="./assets/images/logo.png" class="logo_photographer">
         </header>
         <main id="main" class="margin_left_right">
             <div class="photograph_header">
@@ -60,12 +59,13 @@
         </main>
         <footer></footer>
     </body>
-    <script src="scripts/utils/fetch.js"></script>
+    <script src="./dist/photographer.js"></script>
+    <!-- <script src="scripts/utils/fetch.js"></script>
     <script src="scripts/utils/dom.js"></script>
     <script src="scripts/utils/getUrlParameter.js"></script>
     <script src="scripts/utils/contactForm.js"></script>
     <script src="scripts/factories/mediaFactory.js"></script>
     <script src="scripts/factories/photographerFactory.js"></script>
     <script src="scripts/pages/sharedFunction.js"></script>
-    <script src="scripts/pages/photographer.js"></script>
+    <script src="scripts/pages/photographer.js"></script> -->
 </html>
