+++ src/scripts/factories/photographerFactory.js
