+++ src/scripts/utils/getUrlParameter.js
