+++ src/scss/components/_modal.scss
