+++ src/scss/components/_photograph_header.scss
