+++ src/scss/components/_photographer_cards.scss
