+++ src/scss/components/_photographer_statistic.scss
